#34: Refactoring Dot
