+++ src/main/java/org/cqfn/patternika/visualizer/DotHoleStyle.java
@@ -1,0 +1,37 @@
+package org.cqfn.patternika.visualizer;
+
+import org.cqfn.patternika.ast.Hole;
+
+import java.util.Objects;
+import java.util.function.Consumer;
+
+/**
+ * Hole style for a Dot graph.
+ *
+ * @since 2021/02/16
+ */
+public class DotHoleStyle implements Consumer<StringBuilder> {
+    /** The hole. */
+    private final Hole hole;
+
+    /**
+     * Constructor.
+     *
+     * @param hole the hole.
+     */
+    public DotHoleStyle(final Hole hole) {
+        this.hole = Objects.requireNonNull(hole);
+    }
+
+    /**
+     * Writes the style for a hole to a string builder.
+     *
+     * @param builder the string builder.
+     */
+    @Override
+    public void accept(final StringBuilder builder) {
+        builder.append("style=\"rounded,filled\" color=\"mediumpurple\" fillcolor=\"")
+               .append(hole.getNumber() < 0 ? "thistle1" : "thistle")
+               .append("\" penwidth=2 ");
+    }
+}
